Fixing Bugs and Adding Edit Transactions and Other Features

+++ app/src/main/java/com/bs/sriwilis/data/response/TransactionResponse.kt
@@ -63,7 +63,7 @@ data class TransaksiSampahItemResponse(
 data class TransaksiSampahItem(
 
 	@field:SerializedName("berat")
-	val berat: Int? = null,
+	val berat: Float? = null,
 
 	@field:SerializedName("kategori")
 	val kategori: String? = null,
