+++ app/src/main/java/com/bs/sriwilis/data/room/entity/TransaksiSampahEntity.kt
@@ -17,7 +17,7 @@ data class TransaksiSampahEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val id_keranjang_transaksi: Int?,
     val kategori: String?,
-    val berat: Int?,
+    val berat: Float?,
     val harga: String?,
     val gambar: String?,
     val created_at: String?,
